requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
python-Levenshtein>=0.23.0
scikit-learn>=1.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11